requirements.txt
# Core Backend Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Image Processing
pillow>=10.0.0
mss>=9.0.0
imagehash>=4.3.0

# Utilities
loguru>=0.7.0
python-multipart>=0.0.6
aiofiles>=23.2.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# AI Integration
openai>=1.3.0
anthropic>=0.7.0

# Vector Database & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.24.0
